Fix package after haphazardly rename
Changing the name with a last commit in the last PR did more damage than
good. This properly renames the package to 'swed-17' throughout the
codebase.
