+++ setup.cfg
@@ -1,5 +1,5 @@
 [metadata]
-name = swed-17
+name = swed_17
 version = 0.1
 author = Joachim Meyer
