+++ src/swed_17/rasterize_zone.py
